added some functionality in controllers

+++ src/main/java/com/smuniov/restaurantServiceSystem/service/impl/EmployeeServiceImpl.java
@@ -1,12 +1,16 @@
 package com.smuniov.restaurantServiceSystem.service.impl;
 
+import com.smuniov.restaurantServiceSystem.Exception.BadRequestException;
+import com.smuniov.restaurantServiceSystem.entity.enumeration.Position;
 import com.smuniov.restaurantServiceSystem.entity.users.Employee;
 import com.smuniov.restaurantServiceSystem.repository.EmployeeRepository;
 import com.smuniov.restaurantServiceSystem.service.EmployeeServiceI;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.security.Principal;
 import java.util.List;
 import java.util.Optional;
 
@@ -30,6 +34,11 @@ public class EmployeeServiceImpl implements EmployeeServiceI {
 
     @Override
     public void deleteById(int employeeId) {
+        Principal principal = (Principal) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        String emailLogged = principal.getName();
+        if(employeeRepository.existsByEmail(emailLogged)){
+            throw new BadRequestException("you cannot delete yourself!!!");
+        }
         employeeRepository.deleteById(employeeId);
     }
 
@@ -49,7 +58,7 @@ public class EmployeeServiceImpl implements EmployeeServiceI {
 
     @Override
     public Employee getFree(String position) {
-        return (Employee) employeeRepository.findAllByOrderByLoadFactorAsc().get(0);
+        return (Employee) employeeRepository.findAllByPositionOrderByLoadFactorAsc(Position.valueOf(position)).get(0);
     }
 
     @Override
